Implemented stretch goal for More-OOP
