+++ src/main/java/command/parser/CommandsHandler.java
@@ -1,0 +1,42 @@
+package command.parser;
+
+import exceptions.IllegalCommandException;
+import exceptions.IllegalListException;
+import exceptions.IllegalTaskException;
+import exceptions.IllegalTaskRedoException;
+import task.list.TaskList;
+import ui.UI;
+
+import java.util.ArrayList;
+
+public class CommandsHandler {
+    public static void handleDelete(ArrayList<TaskList> tasks, String[] sentence) {
+        try {
+            DeleteCommand.deleteTask(sentence, tasks);
+        } catch (IllegalCommandException e) {
+            UI.printCommandDoesNotExist();
+        } catch (IllegalTaskException e) {
+            UI.printInvalidTaskPhrase();
+        }
+    }
+
+    public static void handleMarkAsDone(ArrayList<TaskList> tasks, String[] sentence) {
+        try {
+            DoneCommand.markTaskAsDone(sentence, tasks);
+        } catch (IllegalCommandException e) {
+            UI.printCommandDoesNotExist();
+        } catch (IllegalTaskRedoException e) {
+            UI.printTaskAlreadyCompletedPhrase(tasks);
+        } catch (IllegalTaskException e) {
+            UI.printInvalidTaskPhrase();
+        }
+    }
+
+    public static void handleList(ArrayList<TaskList> tasks) {
+        try {
+            ListCommand.printAllLists(tasks);
+        } catch (IllegalListException e) {
+            UI.printEmptyList();
+        }
+    }
+}
